[Enhance] refactor PostCardWideThumbnail
