--- src/components/post-list/card/wide/card-thumbnail/props.d.ts
@@ -1,7 +1,0 @@
-type PostCardWideThumbnailProps = {
-  titleType: string;
-  titleImageUrl: string;
-  titleYoutubeUrl: string;
-};
-
-export default PostCardWideThumbnailProps;
