+++ src/lib/utils/url-parser.ts
@@ -1,14 +1,14 @@
 import {addSizeToImagePath, ImageSize} from './image-size-parser';
 
 export const imageUriHandler = (
-  titleType: string,
-  titleImageUrl: string,
-  titleYoutubeUrl: string,
+  thumbnailType: string,
+  thumbnail: string,
+  youtubeVideoId: string,
   size: ImageSize,
 ) => {
-  if (titleType === 'IMAGE') {
-    return addSizeToImagePath(titleImageUrl, size);
+  if (thumbnailType === 'IMAGE') {
+    return addSizeToImagePath(thumbnail, size);
   } else {
-    return `https://img.youtube.com/vi/${titleYoutubeUrl}/sddefault.jpg`;
+    return `https://img.youtube.com/vi/${youtubeVideoId}/sddefault.jpg`;
   }
 };
